#7 マルチモジュール化対応(domain layer)
