+++ app/build.gradle
@@ -38,7 +38,8 @@ android {
 }
 
 dependencies {
-    implementation project(':data')
+    implementation project(':data') // TODO 消したい
+    implementation project(':domain')
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
     implementation 'androidx.core:core-ktx:1.5.0'
