--- data/src/main/java/com/example/androidapptemplate/data/exception/ApplicationException.kt
@@ -1,3 +1,0 @@
-package com.example.androidapptemplate.data.exception
-
-open class ApplicationException : Exception()
