+++ domain/build.gradle
@@ -1,13 +1,9 @@
 plugins {
-    id 'java-library'
     id 'kotlin'
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_1_7
-    targetCompatibility = JavaVersion.VERSION_1_7
 }
 
 dependencies {
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
+    // javax.inject.Inject 使うために依存追加
+    implementation group: 'javax.inject', name: 'javax.inject', version: '1'
 }
