--- domain/src/main/java/com/example/androidapptemplate/domain/MyClass.kt
@@ -1,4 +1,0 @@
-package com.example.androidapptemplate.domain
-
-class MyClass {
-}
