+++ domain/src/main/java/com/example/androidapptemplate/domain/exception/ApplicationException.kt
@@ -1,0 +1,3 @@
+package com.example.androidapptemplate.domain.exception
+
+open class ApplicationException : Exception()
