+++ domain/src/main/java/com/example/androidapptemplate/domain/exception/BadRequestErrorException.kt
@@ -1,3 +1,3 @@
-package com.example.androidapptemplate.data.exception
+package com.example.androidapptemplate.domain.exception
 
 class BadRequestErrorException(override val message: String) : ApplicationException()
