+++ domain/src/main/java/com/example/androidapptemplate/domain/repository/TriviaRepository.kt
@@ -1,0 +1,6 @@
+package com.example.androidapptemplate.domain.repository
+
+interface TriviaRepository {
+    suspend fun getRandomTrivia(): String
+    suspend fun insertTriviaInfo(text: String)
+}
